Move BaseBackfillCommand to a new figures.management.base module
